@@ -5,6 +5,7 @@ repository: https://github.com/yh-luo/webviewtube
 
 environment:
   sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.16.6"
 
 platforms:
   android:
@@ -13,15 +14,15 @@ platforms:
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^6.0.3
-  webview_flutter: ^4.0.1
-  webview_flutter_android: ^3.7.1
-  webview_flutter_wkwebview: ^3.4.4
+  provider: ^6.1.4
+  webview_flutter: ^4.8.0
+  webview_flutter_android: ^3.16.3
+  webview_flutter_wkwebview: ^3.14.0
 
 dev_dependencies:
-  build_runner: ^2.1.11
-  flutter_lints: ^2.0.1
+  build_runner: ^2.4.9
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
-  mockito: ^5.3.0
-  webview_flutter_platform_interface: ^2.3.1
+  mockito: ^5.4.4
+  webview_flutter_platform_interface: ^2.8.0
